Fix filter core + change model exec to info
